Split out presence methods into separate module
