+++ lib/client/doc.js
@@ -2,6 +2,7 @@ var emitter = require('../emitter');
 var logger = require('../logger');
 var ShareDBError = require('../error');
 var types = require('../types');
+var presenceMethods = require('./presence');
 
 /**
  * A Doc is a client's view on a sharejs document.
@@ -55,6 +56,7 @@ var types = require('../types');
  */
 
 module.exports = Doc;
+Object.assign(Doc.prototype, presenceMethods);
 function Doc(connection, collection, id) {
   emitter.EventEmitter.call(this);
 
@@ -71,42 +73,7 @@ function Doc(connection, collection, id) {
   // If this.presence is falsy (undefined), it means that
   // the enablePresence flag was not passed into the ShareDB constructor,
   // so the presence features should be disabled..
-  this.presence = connection.enablePresence && {
-
-    // The current presence data.
-    // Map of src -> presence data
-    // Local src === ''
-    current: {},
-
-    // The presence objects received from the server.
-    // Map of src -> presence
-    received: {},
-
-    // The minimum amount of time to wait before removing processed presence from this.presence.received.
-    // The processed presence is removed to avoid leaking memory, in case peers keep connecting and disconnecting a lot.
-    // The processed presence is not removed immediately to enable avoiding race conditions, where messages with lower
-    // sequence number arrive after messages with higher sequence numbers.
-    receivedTimeout: 60000,
-
-    // If set to true, then the next time the local presence is sent,
-    // all other clients will be asked to reply with their own presence data.
-    requestReply: true,
-
-    // A list of ops sent by the server. These are needed for transforming presence data,
-    // if we get that presence data for an older version of the document.
-    cachedOps: [],
-
-    // The ops are cached for at least 1 minute by default, which should be lots, considering that the presence
-    // data is supposed to be synced in real-time.
-    cachedOpsTimeout: 60000,
-
-    // The sequence number of the inflight presence request.
-    inflightSeq: 0,
-
-    // Callbacks (or null) for pending and inflight presence requests.
-    pending: null,
-    inflight: null
-  };
+  this.presence = connection.enablePresence && this._initializePresence();
 
   // Array of callbacks or nulls as placeholders
   this.inflightFetch = [];
@@ -1089,284 +1056,5 @@ function callEach(callbacks, err) {
   return called;
 }
 
-// *** Presence
-
-Doc.prototype.submitPresence = function (data, callback) {
-  if (data != null) {
-    if (!this.type) {
-      var doc = this;
-      return process.nextTick(function() {
-        var err = new ShareDBError(4015, 'Cannot submit presence. Document has not been created. ' + doc.collection + '.' + doc.id);
-        if (callback) return callback(err);
-        doc.emit('error', err);
-      });
-    }
-
-    if (!this.type.createPresence || !this.type.transformPresence) {
-      var doc = this;
-      return process.nextTick(function() {
-        var err = new ShareDBError(4027, 'Cannot submit presence. Document\'s type does not support presence. ' + doc.collection + '.' + doc.id);
-        if (callback) return callback(err);
-        doc.emit('error', err);
-      });
-    }
-
-    data = this.type.createPresence(data);
-  }
-
-  if (this._setPresence('', data, true) || this.presence.pending || this.presence.inflight) {
-    if (!this.presence.pending) {
-      this.presence.pending = [];
-    }
-    if (callback) {
-      this.presence.pending.push(callback);
-    }
-
-  } else if (callback) {
-    process.nextTick(callback);
-  }
-
-  process.nextTick(this.flush.bind(this));
-};
-
-Doc.prototype._handlePresence = function(err, presence) {
-  if (!this.subscribed) return;
-
-  var src = presence.src;
-  if (!src) {
-    // Handle the ACK for the presence data we submitted.
-    // this.presence.inflightSeq would not equal presence.seq after a hard rollback,
-    // when all callbacks are flushed with an error.
-    if (this.presence.inflightSeq === presence.seq) {
-      var callbacks = this.presence.inflight;
-      this.presence.inflight = null;
-      this.presence.inflightSeq = 0;
-      var called = callbacks && callEach(callbacks, err);
-      if (err && !called) this.emit('error', err);
-      this.flush();
-      this._emitNothingPending();
-    }
-    return;
-  }
-
-  // This shouldn't happen but check just in case.
-  if (err) return this.emit('error', err);
-
-  if (presence.r && !this.presence.pending) {
-    // Another client requested us to share our current presence data
-    this.presence.pending = [];
-    this.flush();
-  }
-
-  // Ignore older messages which arrived out of order
-  if (
-    this.presence.received[src] && (
-      this.presence.received[src].seq > presence.seq ||
-      (this.presence.received[src].seq === presence.seq && presence.v != null)
-    )
-  ) return;
-
-  this.presence.received[src] = presence;
-
-  if (presence.v == null) {
-    // null version should happen only when the server automatically sends
-    // null presence for an unsubscribed client
-    presence.processedAt = Date.now();
-    return this._setPresence(src, null, true);
-  }
-
-  // Get missing ops first, if necessary
-  if (this.version == null || this.version < presence.v) return this.fetch();
-
-  this._processReceivedPresence(src, true);
-};
-
-// If emit is true and presence has changed, emits a presence event.
-// Returns true, if presence has changed for src. Otherwise false.
-Doc.prototype._processReceivedPresence = function(src, emit) {
-  if (!src) return false;
-  var presence = this.presence.received[src];
-  if (!presence) return false;
-
-  if (presence.processedAt != null) {
-    if (Date.now() >= presence.processedAt + this.presence.receivedTimeout) {
-      // Remove old received and processed presence.
-      delete this.presence.received[src];
-    }
-    return false;
-  }
-
-  if (this.version == null || this.version < presence.v) {
-    // keep waiting for the missing snapshot or ops.
-    return false;
-  }
-
-  if (presence.p == null) {
-    // Remove presence data as requested.
-    presence.processedAt = Date.now();
-    return this._setPresence(src, null, emit);
-  }
-
-  if (!this.type || !this.type.createPresence || !this.type.transformPresence) {
-    // Remove presence data because the document is not created or its type does not support presence
-    presence.processedAt = Date.now();
-    return this._setPresence(src, null, emit);
-  }
-
-  if (this.inflightOp && this.inflightOp.op == null) {
-    // Remove presence data because presence.received can be transformed only against "op", not "create" nor "del"
-    presence.processedAt = Date.now();
-    return this._setPresence(src, null, emit);
-  }
-
-  for (var i = 0; i < this.pendingOps.length; i++) {
-    if (this.pendingOps[i].op == null) {
-      // Remove presence data because presence.received can be transformed only against "op", not "create" nor "del"
-      presence.processedAt = Date.now();
-      return this._setPresence(src, null, emit);
-    }
-  }
-
-  var startIndex = this.presence.cachedOps.length - (this.version - presence.v);
-  if (startIndex < 0) {
-    // Remove presence data because we can't transform presence.received
-    presence.processedAt = Date.now();
-    return this._setPresence(src, null, emit);
-  }
-
-  for (var i = startIndex; i < this.presence.cachedOps.length; i++) {
-    if (this.presence.cachedOps[i].op == null) {
-      // Remove presence data because presence.received can be transformed only against "op", not "create" nor "del"
-      presence.processedAt = Date.now();
-      return this._setPresence(src, null, emit);
-    }
-  }
-
-  // Make sure the format of the data is correct
-  var data = this.type.createPresence(presence.p);
-
-  // Transform against past ops
-  for (var i = startIndex; i < this.presence.cachedOps.length; i++) {
-    var op = this.presence.cachedOps[i];
-    data = this.type.transformPresence(data, op.op, presence.src === op.src);
-  }
-
-  // Transform against pending ops
-  if (this.inflightOp) {
-    data = this.type.transformPresence(data, this.inflightOp.op, false);
-  }
-
-  for (var i = 0; i < this.pendingOps.length; i++) {
-    data = this.type.transformPresence(data, this.pendingOps[i].op, false);
-  }
-
-  // Set presence data
-  presence.processedAt = Date.now();
-  return this._setPresence(src, data, emit);
-};
-
-Doc.prototype._processAllReceivedPresence = function() {
-  if (!this.presence) return;
-  var srcList = Object.keys(this.presence.received);
-  var changedSrcList = [];
-  for (var i = 0; i < srcList.length; i++) {
-    var src = srcList[i];
-    if (this._processReceivedPresence(src)) {
-      changedSrcList.push(src);
-    }
-  }
-  this._emitPresence(changedSrcList, true);
-};
-
-Doc.prototype._transformPresence = function(src, op) {
-  var presenceData = this.presence.current[src];
-  if (op.op != null) {
-    var isOwnOperation = src === (op.src || '');
-    presenceData = this.type.transformPresence(presenceData, op.op, isOwnOperation);
-  } else {
-    presenceData = null;
-  }
-  return this._setPresence(src, presenceData);
-};
-
-Doc.prototype._transformAllPresence = function(op) {
-  if (!this.presence) return;
-  var srcList = Object.keys(this.presence.current);
-  var changedSrcList = [];
-  for (var i = 0; i < srcList.length; i++) {
-    var src = srcList[i];
-    if (this._transformPresence(src, op)) {
-      changedSrcList.push(src);
-    }
-  }
-  this._emitPresence(changedSrcList, false);
-};
-
-Doc.prototype._pausePresence = function() {
-  if (!this.presence) return;
-
-  if (this.presence.inflight) {
-    this.presence.pending = this.presence.pending
-      ? this.presence.inflight.concat(this.presence.pending)
-      : this.presence.inflight;
-    this.presence.inflight = null;
-    this.presence.inflightSeq = 0;
-  } else if (!this.presence.pending && this.presence.current[''] != null) {
-    this.presence.pending = [];
-  }
-  this.presence.received = {};
-  this.presence.requestReply = true;
-  var srcList = Object.keys(this.presence.current);
-  var changedSrcList = [];
-  for (var i = 0; i < srcList.length; i++) {
-    var src = srcList[i];
-    if (src && this._setPresence(src, null)) {
-      changedSrcList.push(src);
-    }
-  }
-  this._emitPresence(changedSrcList, false);
-};
-
-// If emit is true and presence has changed, emits a presence event.
-// Returns true, if presence has changed. Otherwise false.
-Doc.prototype._setPresence = function(src, data, emit) {
-  if (data == null) {
-    if (this.presence.current[src] == null) return false;
-    delete this.presence.current[src];
-  } else {
-    var isPresenceEqual =
-      this.presence.current[src] === data ||
-      (this.type.comparePresence && this.type.comparePresence(this.presence.current[src], data));
-    if (isPresenceEqual) return false;
-    this.presence.current[src] = data;
-  }
-  if (emit) this._emitPresence([ src ], true);
-  return true;
-};
-
-Doc.prototype._emitPresence = function(srcList, submitted) {
-  if (srcList && srcList.length > 0) {
-    var doc = this;
-    process.nextTick(function() {
-      doc.emit('presence', srcList, submitted);
-    });
-  }
-};
-
-Doc.prototype._cacheOp = function(op) {
-  if (!this.presence) return;
-  // Remove the old ops.
-  var oldOpTime = Date.now() - this.presence.cachedOpsTimeout;
-  var i;
-  for (i = 0; i < this.presence.cachedOps.length; i++) {
-    if (this.presence.cachedOps[i].time >= oldOpTime) {
-      break;
-    }
-  }
-  if (i > 0) {
-    this.presence.cachedOps.splice(0, i);
-  }
-
-  // Cache the new op.
-  this.presence.cachedOps.push(op);
-};
+// Expose callEach to presence methods via Doc prototype.
+Doc.prototype._callEach = callEach;
